refactor(cache): move `get_or_populate_cache()` to `cache` module
